requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-decouple
alembic
aiosqlite